Resolved conflicts raised in first review of pull request #62
