+++ src/main/java/Kaji.java
@@ -4,7 +4,6 @@ import exception.IncorrectAccessLevelException;
 import exception.InvalidInputException;
 import exception.InvalidFileFormatException;
 import manager.admin.Admin;
-import manager.admin.ModuleList;
 import manager.chapter.CardList;
 import parser.Parser;
 import storage.Storage;
@@ -25,7 +24,7 @@ public class Kaji {
         cards = new CardList();
         storage = new Storage(filePath);
         try {
-            Admin admin = new Admin(new ModuleList(storage.loadModule()));
+            Admin admin = new Admin(storage.loadModule());
             access = new Access(admin);
         } catch (FileNotFoundException e) {
             storage.createAdmin();
